+++ packages/website/src/examples/ColorBar/ColorBarPartial.example.tsx
@@ -1,0 +1,6 @@
+import {ColorBar} from '@coveord/plasma-react';
+import * as React from 'react';
+
+export default () => (
+    <ColorBar widthPerColor={{'#1372ec': 20, '#ffe300': 5, '#f05245': 10, '#1cebcf': 5, '#7d458f': 20}} />
+);