+++ packages/website/src/examples/FlatSelect/FlatSelectOptionPicker.example.tsx
@@ -1,0 +1,23 @@
+import {FlatSelectConnected} from '@coveord/plasma-react';
+import * as React from 'react';
+
+export default () => (
+    <FlatSelectConnected
+        id="flat-select-option-picker"
+        options={[
+            {
+                id: 'item-option-picker-1',
+                option: {content: 'Option 1'},
+            },
+            {
+                id: 'item-option-picker-2',
+                option: {content: 'Option 2'},
+            },
+            {
+                id: 'item-option-picker-3',
+                option: {content: 'Option 3'},
+            },
+        ]}
+        optionPicker
+    />
+);