+++ packages/website/src/examples/JSONEditor/InError.example.tsx
@@ -1,0 +1,12 @@
+import * as React from 'react';
+import {JSONEditorConnected} from '@coveord/plasma-react';
+
+const invalidJSON = '{hello: world}';
+
+export default () => (
+    <JSONEditorConnected
+        id="example-3"
+        defaultValue={invalidJSON}
+        errorMessage="Custom error message when JSON is invalid"
+    />
+);