+++ packages/website/src/examples/ListBox/loading.example.tsx
@@ -1,0 +1,4 @@
+import {ListBox} from '@coveord/plasma-react';
+import * as React from 'react';
+
+export default () => <ListBox isLoading />;