+++ packages/website/src/examples/SyncFeedback/SyncFeedback.example.tsx
@@ -1,0 +1,33 @@
+import {SyncFeedback, SyncFeedbackState} from '@coveord/plasma-react';
+import * as React from 'react';
+
+export default () => (
+    <table className="table">
+        <tr>
+            <td>NONE</td>
+            <td>
+                <SyncFeedback state={SyncFeedbackState.NONE} />
+            </td>
+        </tr>
+        <tr>
+            <td>SYNCING</td>
+            <td>
+                <SyncFeedback state={SyncFeedbackState.SYNCING} />
+            </td>
+        </tr>
+
+        <tr>
+            <td>SUCCESS</td>
+            <td>
+                <SyncFeedback state={SyncFeedbackState.SUCCESS} />
+            </td>
+        </tr>
+
+        <tr>
+            <td>ERROR</td>
+            <td>
+                <SyncFeedback state={SyncFeedbackState.ERROR} />
+            </td>
+        </tr>
+    </table>
+);