+++ packages/website/src/examples/multiline-box/MultilineBox.example.tsx
@@ -1,0 +1,40 @@
+import * as React from 'react';
+import {
+    FormProvider,
+    MultilineBox,
+    IMultilineSingleBoxProps,
+    IMultilineParentProps,
+    TextInput,
+} from '@coveord/plasma-react';
+
+interface MyData {
+    name: string;
+}
+
+export default () => (
+    <FormProvider>
+        <MultilineBox<MyData>
+            id="multiline-box-id"
+            data={[{name: 'Poire'}, {name: 'Apple'}]}
+            renderBody={(data: Array<IMultilineSingleBoxProps<MyData>>, defaultProps: IMultilineParentProps) =>
+                Object.values(data).map(({id, isLast, props}: IMultilineSingleBoxProps<MyData>) => (
+                    <TextInput
+                        id={id}
+                        key={id}
+                        required={!isLast}
+                        showValidationOnBlur
+                        type="text"
+                        label="Label"
+                        defaultValue={props.name}
+                        onBlur={(evt: React.FocusEvent<HTMLInputElement>) => {
+                            if (evt.target.value !== '' && isLast) {
+                                defaultProps.addNewBox();
+                            }
+                        }}
+                    />
+                ))
+            }
+            defaultProps={{name: ''}}
+        />
+    </FormProvider>
+);